refactor: add mixins and import types
