+++ src/pages/RepoListPage/RepoListPage.module.scss
@@ -1,3 +1,5 @@
+@use '@/styles/mixins.scss' as *;
+
 .repo-list {
   flex: 1;
 
@@ -8,15 +10,11 @@
   }
 
   &__title {
-    display: flex;
-    justify-content: center;
-    align-items: center;
+    @include flex-center;
   }
 
   &__loader {
-    display: flex;
-    justify-content: center;
-    align-items: center;
+    @include flex-center;
   }
 
   &__search {
