+++ src/pages/RepoListPage/components/RepoCard/RepoCard.module.scss
@@ -1,3 +1,5 @@
+@use '@/styles/mixins.scss' as *;
+
 .repo-card {
   background: #fff;
   border: 1px solid #e5e7eb;
@@ -34,8 +36,7 @@
   }
 
   &__icon {
-    width: 16px;
-    height: 16px;
+    @include square(16px);
   }
 
   &__stars {
